refactor(bunderstack): enhance type safety and code clarity across modules
- Simplified type handling in access, auth, and storage modules by removing unnecessary type assertions and improving type definitions.
- Introduced utility functions to validate object shapes, enhancing type safety in various functions.
- Cleaned up code in CRUD and realtime modules for better readability and maintainability.
- Updated cursor parsing logic to utilize new type guards for improved validation.

+++ packages/bunderstack/src/typeid.ts
@@ -62,7 +62,7 @@ function bytesToUuid(bytes: Uint8Array): string {
 export function generate<P extends string>(prefix: P): TypeId<P> {
   if (!isValidPrefix(prefix))
     throw new Error(`Invalid typeid prefix: "${prefix}"`)
-  const bytes = Bun.randomUUIDv7('buffer') as Uint8Array
+  const bytes = Bun.randomUUIDv7('buffer')
   return `${prefix}_${encode(bytes)}` as TypeId<P>
 }
